фикс проблем со статами

+++ src/server/game/Entities/Player/Player.h
@@ -297,9 +297,13 @@ typedef std::list<PlayerCreateInfoItem> PlayerCreateInfoItems;
 
 struct PlayerLevelInfo
 {
-    PlayerLevelInfo() { for (uint8 i=0; i < MAX_STATS; ++i) stats[i] = 0; }
+    PlayerLevelInfo()
+    {
+        for (uint8 i = 0; i < MAX_STATS; ++i)
+            stats[i] = 0;
+    }
 
-    uint8 stats[MAX_STATS];
+    uint16 stats[MAX_STATS];
 };
 
 typedef std::list<uint32> PlayerCreateInfoSpells;
